Move SiteMetadata logic to slice-sites
