+++ packages/playground/website/src/lib/state/opfs/opfs-site-storage.ts
@@ -6,7 +6,7 @@
  */
 
 import metadataWorkerUrl from './opfs-site-storage-worker-for-safari?worker&url';
-import type { SiteMetadata } from '../../site-metadata';
+import type { SiteMetadata } from '../redux/slice-sites';
 import type { SiteInfo } from '../redux/slice-sites';
 import { joinPaths } from '@php-wasm/util';
 import { logger } from '@php-wasm/logger';
